Compile issues with Clang
Fix compile issues found when compiling with clang -std=c++11

This commit fixes errors, several warnings still exist.

Details: Clang v 3.5
         openmpi-1.8.1
         Trilinos option:     -D Trilinos_ENABLE_CXX11:BOOL=ON

[gahanse@cee-compute011 ~]$ module list
Currently Loaded Modulefiles:
  1) sierra-compiler/clang/3.5    3) viz
  5) sierra-valgrind/3.9.0
  2) sierra-mkl/15.0-2015.0.090   4) sierra-git/2.0.0

+++ src/disc/stk/Albany_STKDiscretization.hpp
@@ -349,7 +349,7 @@ namespace Albany {
     Teuchos::RCP<Piro::MLRigidBodyModes> rigidBodyModes;
 
     int netCDFp;
-    int netCDFOutputRequest;
+    size_t netCDFOutputRequest;
     std::vector<int> varSolns;
     Albany::WorksetArray<Teuchos::ArrayRCP<std::vector<interp> > >::type interpolateData;
